% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_df.R
\name{center_df}
\alias{center_df}
\title{Center numeric variables in a data.frame returning a data.frame}
\usage{
center_df(x, by = mean)
}
\arguments{
\item{x}{A data.frame as input}

\item{by}{A character vector, numeric vector, or data.frame with scaling information}
}
\value{
a data.frame of the same dimensions as the input with numeric
  columns centered and with additional attributes describing the
  scaling carried out.
}
\description{
\code{center_df} returns a data.frame where each of the columns
is centered by a constant.
}
\examples{
df <- data.frame(x=1:5,y=6:10,z=factor(letters[1:5]))
\dontrun{
scale(df, center = TRUE, scale = FALSE) # an error because of factor
}
centered_df <- center_df(df)
attributes(centered_df)
}
\seealso{
\code{\link[base]{scale}}, \code{\link[arm]{rescale}}, and
  \code{\link[arm]{standardize}}.
  \href{https://stevencarlislewalker.wordpress.com/2012/09/24/a-simpler-version-of-the-r-scale-command/}{Here's a different reason} to replace \code{scale()}
}

