% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_df.R
\name{scale_df}
\alias{scale_df}
\title{Scale a data.frame returning a data.frame}
\usage{
scale_df(x, by = sd)
}
\arguments{
\item{x}{A data.frame as input}

\item{by}{A character vector, numeric vector, or data.frame with scaling
information}
}
\value{
a data.frame of the same dimensions as the input with numeric columns
  scaled and with additional attributes describing the scaling carried out.
}
\description{
\code{scale_df} returns a data.frame where each of the columns is scaled by a
constant.
}
\examples{
df <- data.frame(x=1:5,y=6:10,z=factor(letters[1:5]))
\dontrun{
scale(df, center = FALSE, scale = apply(x, 2, sd, na.rm = TRUE)) # an error because of factor
}
scaled_df <- scale_df(df)
attributes(scaled_df)
}
\seealso{
\code{\link[base]{scale}}, \code{\link[arm]{rescale}}, and
  \code{\link[arm]{standardize}}.
  \href{https://stevencarlislewalker.wordpress.com/2012/09/24/a-simpler-version-of-the-r-scale-command/}{Here's a different reason} to replace \code{scale()}
}

