% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_df.R
\name{uncenter_df}
\alias{uncenter_df}
\title{Uncenter numeric variables in a data.frame returning a data.frame}
\usage{
uncenter_df(x, by = NULL)
}
\arguments{
\item{x}{A data.frame as input, possibly including centering information}

\item{by}{A numeric vector}
}
\value{
a data.frame of the same dimensions as the input with numeric
  columns uncentered. Centering attributes are removed.
}
\description{
\code{uncenter_df} returns a data.frame where each of the columns
has been unentered by a constant.
}
\examples{
df <- data.frame(x=1:5,y=6:10,z=factor(letters[1:5]))
\dontrun{
scale(df, center = TRUE, scale = FALSE) # an error because of factor
}
centered_df <- center_df(df)
attributes(centered_df)
uncentered_df <- uncenter_df(centered_df)
all.equal(uncentered_df, df)
}
\seealso{
\code{\link[base]{scale}}, \code{\link[arm]{rescale}}, and
  \code{\link[arm]{standardize}}.
  \href{https://stevencarlislewalker.wordpress.com/2012/09/24/a-simpler-version-of-the-r-scale-command/}{Here's a different reason} to replace \code{scale()}
}

