% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaler-package.R
\docType{package}
\name{scaler-package}
\alias{scaler}
\alias{scaler-package}
\title{scaler: Tools to Process Health Measurement Scales}
\description{
A set of tools to deal with health measurement scales.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gipso.github.io/scaler/}
  \item \url{https://github.com/gipso/scaler/}
  \item Report bugs at \url{https://github.com/gipso/scaler/issues/}
}

}
\author{
\strong{Maintainer}: Daniel Vartanian \email{danvartan@gmail.com} (\href{https://orcid.org/0000-0001-7782-759X}{ORCID}) [conceptor, copyright holder]

Authors:
\itemize{
  \item Mario Pedrazzoli \email{pedrazzo@usp.br} (\href{https://orcid.org/0000-0002-5257-591X}{ORCID}) [scientific advisor]
}

Other contributors:
\itemize{
  \item Interdisciplinary Sleep Research Group (GIPSO) [funder]
  \item University of Sao Paulo (USP) [funder]
}

}
\keyword{internal}
