% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaler-package.R
\docType{package}
\name{scaler-package}
\alias{scaler}
\alias{scaler-package}
\title{scaler: Tools to Process Health Measurement Scales}
\description{
A set of tools to process health measurement scales.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://giperbio.github.io/scaler/}
  \item \url{https://github.com/giperbio/scaler/}
  \item Report bugs at \url{https://github.com/giperbio/scaler/issues/}
}

}
\author{
\strong{Maintainer}: Daniel Vartanian \email{danvartan@gmail.com} (\href{https://orcid.org/0000-0001-7782-759X}{ORCID}) [conceptor, copyright holder]

}
\keyword{internal}
